+++ src/settings/StorageSettings.js
@@ -2,7 +2,7 @@ import { sortBy } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { injectIntl } from 'react-intl';
-import { useStripes, stripesConnect } from '@folio/stripes/core';
+import { stripesConnect } from '@folio/stripes/core';
 import { EntryManager } from '../smart-components';
 
 import StorageDetail from './StorageDetail';
@@ -16,10 +16,6 @@ const PERMS = {
 
 const StorageSettings = (props) => {
   const { mutator, resources, intl } = props;
-  const stripes = useStripes();
-
-  // When this is disabled <EntryManager> displays an Edit button, which there is no way to remove. See STSMACOM-764
-  const enableDetailsActionMenu = stripes.hasPerm(PERMS.put) || stripes.hasPerm(PERMS.post) || stripes.hasPerm(PERMS.delete);
 
   return (
     <EntryManager
@@ -33,7 +29,7 @@ const StorageSettings = (props) => {
       entryFormComponent={StorageForm}
       nameKey="name"
       permissions={PERMS}
-      enableDetailsActionMenu={enableDetailsActionMenu}
+      enableDetailsActionMenu
       parseInitialValues={values => {
         if (!values.json) return values;
         return { ...values, json: JSON.stringify(values.json, null, 2) };